Add two options to examples/spmm:   - -mm /path/to/a/MatrixMarket/sparse/matrix   - -rmat N[:E[:a[:b[:c[:d]]]]] use a random graph generator     in Boost to build the connectivity matrix of that graph
If either -mm or -rmat are used, we compute C = A * A, if A is
square, or C = A * A^T, if A is not square, where A
is the matrix provided / generated. Some rudimentary
statistics about the matrices (size and number of non zero),
and the time to compute the product.

If neither -mm or -rmat are specified, the example continues to
use the previous behavior.