key embedded in the parsec_task_t object
When a parsec_task_t object is created, the key fo the task
is packed into the ttg_key member in the object. When the task
is selected for migration, the key is unpacked from the same
object.

Now all theinformation to help migrate a task is available
just from the parsec_task_t object.