task migration implemented using task_class member migrate_task
When parsec decides a task has to be migrated it calls the migrate_task
member (this is a function pointer) of a task class. Depdending
on the type of class the function pointer points to function
that migrated the task, or a dummy function which does not do
anything.

if constexpr ( !ttg::meta::is_empty_tuple_v<input_refs_tuple_type> )
	self.migrate_task = migrate_task_to_dst;
else
        self.migrate_task = migrate_task_dummy;

Now the decision to migrate lies with parsec. While the actual migration
is done by TTG functions.