function edge_number() returns the input edge number the aggregator task to send the data to.
We use a map object

std::map<Key, std::atomic<int>> key_edge_map

to manage all the keys of an aggregator task. a integer
value is associated with each key and integer value is
incremented each time an octant task calls the edge_number().
This will ensure that each task that sends data to an
aggregator task sends it to the different input edge.