refinement decision logic changed.
The refinement decision of each octant is mapped using
std::map<Key, Octant_Aggregator_Data> refinement;

The octant task calls refinement_decision with key of the octant task as
the key of the map. The value of the map is an object of type Octant_Aggregator_Data.
Once information is aggregated at the root aggregator, refinement_parse() will finalise,
the refinement level of each octant in the next timestep. Based on  the decisions made
by refinement_parse() the tasks are refined.